backend/requirements.txt
# Add this to your requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.1
langgraph 
langchain 
langchain-groq 
langchain-community
faiss-cpu 
python-dotenv 
pydantic 
pandas 
numpy
uvicorn 
fastapi 
python-multipart
langgraph-cli[inmem]
langgraph[cli]
fastapi
uvicorn
sentence-transformers